better code organization, allow loading pickled sparse matrices
