+++ datasets.py
@@ -1,11 +1,13 @@
 import numpy as np
 import scipy.sparse as sparse
+import pickle
+from collections import defaultdict
 
 
 print("Datasets Available")
 print("-" * 20)
 # ################### Binary Tree
-depth = 15 
+depth = 10 
 N = sum(2 ** i for i in range(depth))
 mat = np.zeros((N, N))
 for i in range(N):
@@ -35,16 +37,22 @@ print("Quad Tree        : quad_tree")
 print("-" * 20)
 
 
+
+def get_pickle_dataset(filename):
+    return pickle.load(open(filename, 'rb'))
+
 datasets = {
     'binary_tree': binary_tree,
-    'quad_tree': quad_tree,
+    'quad_tree':  quad_tree
 }
 
+print(' '.join(datasets.keys()))
 
-import pickle
+def get_dataset(dataset_name_or_filename):
+    print("using {} dataset".format(dataset_name_or_filename))
+    if dataset_name_or_filename in datasets.keys():
+        print("using builtin dataset")
+        return datasets[dataset_name_or_filename]
+    else:
+        return get_pickle_dataset(dataset_name_or_filename)
 
-try:
-    call_graph = pickle.load(open('call_graph.pkl', 'rb'))
-    datasets['call_graph'] = call_graph.mat
-except:
-    pass
